Fixed Query.py error, Moved converting of image to data into ReadGraph function in IR_Functions
