+++ IR_Functions.py
@@ -3,10 +3,25 @@
 """
 #---------------------------------Imports--------------------------------------
 import PyPDF2
+from PIL import Image, ImageTk
 import sys
 import warnings
 import os
 from os import path
+from shutil import copyfile
+import PyPDF2
+import sys
+import sqlite3
+import warnings
+import os
+from os import path
+from PIL import Image, ImageTk
+from math import log
+from shutil import copyfile
+
+import multiprocessing as mp
+
+import time
 warnings.filterwarnings("ignore")
 #------------------------------------------------------------------------------
 
@@ -83,6 +98,91 @@ def PullImages(filename):
     images=recurse(p, page0)
 
     return images
+
+def ReadGraph(image):
+    """ Crop the image """
+    img = Image.open(image)
+    imgdata=list(img.getdata())#the pixels from the image
+
+    #this with and height is standard for all IR samples
+    Width=1024
+    Height=768
+
+    #the area of each image that we want (the graph)
+            #(left,right,top,bottom)
+    targetRect=(113,978,29,724)
+
+    def cropRect(source,rect):#copies pixels from the source image within the targetRect
+        left,right,top,bottom=rect
+        newImg=[]
+        for y in range(top,bottom+1):
+            for x in range(left,right+1):
+                newImg+=[source[y*Width+x]]
+        return newImg
+
+    #the graph cut out of the larger image
+    graph=cropRect(imgdata,targetRect)
+
+    #width and height of out cropped graph
+    Width=targetRect[1]-targetRect[0]+1
+    Height=targetRect[3]-targetRect[2]+1
+
+    '''
+    Create graphData list by reading pixels from graph
+        -each entry in data is the range over wich each
+         column has black pixels
+    Scale to x and y units
+    Save data to file
+    '''
+
+    graphData=[]#to be filled with values from graph
+
+    def pix(x,y):#checks if the pixel at x,y is black
+        r,g,b=graph[y*Width+x]
+        if r+g+b>=100:
+            return False#not black
+        else:
+            return True#black
+
+    #For each x get the y range over which the graph has black pixels
+    # or None if the graph is empty at that x value
+    for x in range(0,Width):
+        graphData+=[None]
+        foundPix=False#have you found a pixel while looping through the column
+        for y in range(0,Height):
+            p=pix(x,y)#is the pixel black
+            if p and not foundPix:
+                #record the first black pixels y value
+                foundPix=True
+                maxVal=y
+            elif not p and foundPix:
+                #record the last black pixels y value
+                minVal=y
+                graphData[-1]=(minVal,maxVal)#write these values to data
+                break#next x
+
+    #the range each axis of the graph covers
+    yMin=1.02
+    yMax=-0.05
+    yRange=yMax-yMin
+    xMin=200
+    xMax=4100
+    xRange=xMax-xMin
+
+    #convert graph x,y into scientific x,y
+    def convertx(x):
+        return xMin+xRange*(x/Width)
+    def converty(y):
+        return yMin+yRange*(y/Height)
+
+    data=[]#final value written to file
+    #convert graph into datapoints
+    for x in range(len(graphData)):
+        #Points in format x,y
+        if graphData[x]:
+            data+=[(convertx(x),
+                    converty(graphData[x][0]),converty(graphData[x][1]))]
+    return(data)
 
 def Cumulative(l):
     l=['x']+l[:]+['x']
